requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database - SQLAlchemy with async support
sqlalchemy[asyncio]==2.0.25
aiomysql==0.2.0
aiosqlite==0.19.0
pymysql==1.1.0

# Scraping
selenium==4.16.0
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
lxml==5.1.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development
black==23.12.1
isort==5.13.2
cryptography>=42.0.0

# AI Integration
google-generativeai>=0.3.0

# Caching
redis>=5.0.0
